requirements.txt
cupy
numpy
tflearn
matplotlib
scipy
pillow
tensorflow
python-mnist
pygame
opencv-python
requests
python-mnist